update flutter uikit to 0.1.1

+++ Flutter/IMSDK/tim_ui_kit/lib/business_logic/view_models/tui_block_list_view_model.dart
@@ -1,13 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:tencent_im_sdk_plugin/models/v2_tim_friend_operation_result.dart';
+import 'package:tim_ui_kit/business_logic/life_cycle/block_list_life_cycle.dart';
 import 'package:tim_ui_kit/data_services/friendShip/friendship_services.dart';
 import 'package:tim_ui_kit/data_services/services_locatar.dart';
 import 'package:tim_ui_kit/tim_ui_kit.dart';
 
-class TUIBlackListViewModel extends ChangeNotifier {
+class TUIBlockListViewModel extends ChangeNotifier {
   List<V2TimFriendInfo> _blackList = [];
   final FriendshipServices _friendshipServices =
       serviceLocator<FriendshipServices>();
+  late BlockListLifeCycle? _lifeCycle;
+
+  set lifeCycle(BlockListLifeCycle? value) {
+    _lifeCycle = value;
+  }
 
   List<V2TimFriendInfo> get blackList {
     return _blackList;
@@ -21,6 +26,10 @@ class TUIBlackListViewModel extends ChangeNotifier {
 
   Future<List<V2TimFriendOperationResult>?> deleteFromBlackList(
       List<String> userIDList) async {
+    if (_lifeCycle?.shouldDeleteFromBlockList != null &&
+        await _lifeCycle!.shouldDeleteFromBlockList(userIDList) == false) {
+      return null;
+    }
     final res =
         await _friendshipServices.deleteFromBlackList(userIDList: userIDList);
     if (res != null) {
